requirements.txt
fastapi>=0.95.0
pydantic>=2.0.0
uvicorn[standard]>=0.18.0
motor>=3.1.1
dnspython>=2.3.0
python-dotenv>=1.0.0